Related: tdf#54053 PDF export: add UNO API to customize the watermark font size
The font height of the watermark text is currently automatic: we start
with a value based on the page size and then decrease it till the text
fits.

The problem is that sometimes you want a smaller value, but specifying
this was not possible.

Fix the problem by adding a new "WatermarkFontHeight" PDF export filter
option to specify the font size explicitly.

Example cmdline usage:

soffice --convert-to pdf:writer_pdf_Export:'{"Watermark":{"type":"string","value":"draft"}, "WatermarkFontHeight":{"type":"long","value":"100"}}' test.odt

Change-Id: Iceab943b7a995badfcdc6b12ecc9264e39e4a3a2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/143135
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins