sd: better detection of user defined cell styles
Cell style might be used for several table roles
at once, or even in several table styles (esp.
after new/clone actions). But when we're editing a
style, we want just a single place to be changed.

The initial solution was to check if the style is
user defined. The problem is that if we change the
built-in styles set, then documents created with
the old styles will have their once built-in styles
turned into user defined. In such case we might
still attempt to edit a shared style, which is
unintended.

So try harder, and check also that the style isn't
used twice and that there is no other styles
inheriting from it.

Change-Id: I67d8c61b15161c9d839e97f7a31a1349a7e2613c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/143288
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>