tdf#142147 saveas: encrypted docs should default to use password
If a document is currently encrypted (pOldEncryptionDataItem),
when the user saves as a different name, the save dialog should
BY DEFAULT have the "Save with password" option enabled.

That makes sense, right? If the document was secure enough
to require a password, then any copies of it should
AT MINIMUM suggest a password.

Anyway, it worked that way prior to LO 7.0's
e0cced9d4c94324e834e46d807469a0cd6c1f738.

The concern was that the same password would be applied
silently. That might have been true with 7.0.
However, in my current testing that never happened.
It always prompted the user to create a new password.
And that is exactly what I would want on a SAVE-AS,
whether it be from a template or a file I was copying.

I would hope that this has no bad implications for headless RT.
After all, this is just a "suggestion" passed to guisaveas.cxx.
I did not find any parameters to provide a password
via commandline anyway.

Change-Id: Id3535d883cd5d2d1427055f7a21f6ae87e0d5324
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/143250
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>