tdf#152031,tdf#152085 minimize impact to table layout.
Revert two commits[1][2] that are related to fixing tdf#135991
( but keep the unit tests) and rework the solution.

Cases related to IsNeighborFrame() imply either IsCellFrame()
( a table cell ), or IsColumnFrame() ( a column inside a
section ).  Original patch[2] affected both of tables and
sections.  Try to limit the affected case to the section
so it does not affect table layout for tdf#152031.

[1]
1bf82b26aea3a403920a64cdfcb4671c947c7a01
Author: Mark Hung <marklm9@gmail.com>
Date:   Fri Sep 9 16:09:47 2022 +0800

    tdf#150642 fix table layout with vertical writing.

[2]
1e21902106cbe57658bed03ed24d4d0863685cfd
Author: Mark Hung <marklm9@gmail.com>
Date:   Sun Aug 14 15:23:14 2022 +0800

    tdf#135991 fix unexpected hidden RTL sections.

Change-Id: If50e2b50b47352dad8eafebd6cce0af0530abefa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/143317
Tested-by: Jenkins
Reviewed-by: Mark Hung <marklh9@gmail.com>