tdf#124601 sw: fix misplaced images in table cells
The bugdoc had a table with anchored images which were
FollowTextFlow=true and WrapThrough=true. This used to still make the
table cell larger, but instead the images were overlapping due to the
not increased cell height.

The problem happens since commit
489eef894e7034873ad262f9dfca554022db1b09 (tdf#124601 sw FollowTextFlow:
fix vert pos of objects outside the current cell, 2019-10-01), which
wanted to ignore anchored images in a similar situation.

Fix the problem by making the "ignore image for cell height calculation"
condition more strict: if we also require wrapping in background, then
the old bugdoc keeps working but this fixes the new bugdoc.

Also change the font name in lcl_PaintReplacement() to a font that we
bundle, otherwise the testcase would abort in
psp::PrintFontManager::Substitute() for a document that doesn't specify
any font name explicitly.

Change-Id: Ie512aa4ffd368774825c5243ecd6bce7f09e1442
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/143472
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>