tdf#86568 Move the scrollbar to the EditWindow in Basic IDE
Currently, the Basic IDE has a shared horizontal scrollbar used both for the code editor window and the dialog editor window.

However, the problem is that this causes the horizontal scrollbar to be placed at an unintuitive place when the code editor is being used. To avoid this problem the code editor already has its own vertical scrollbar.

This patch creates a separate horizontal scrollbar for the code editor window and places it underneath the editor window to make it clearer what this scrollbar is meant to scroll.

Change-Id: Ib873c004db902f0f729fd512c51effd8f5e38a9d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/143580
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>