tdf#53023 - Remove last empty paragraph from auto text
Remove the last empty paragraph before inserting an auto text.
The carriage return at the end of it is appended in
SwXMLTextBlockParContext::~SwXMLTextBlockParContext()
which can't be removed without introducing side effects.

In order to compensate for the change, a paragraph has been
added to the existing only text auto texts.

Patch was reverted due to a hang of the ui test. It seems that
selecting an entire text using mkPropertyValues({"KEYCODE":"CTRL+A"}))
does not work anymore.

Change-Id: I08f612d50f496c8d7f891c773ddedf27c3d80864
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/143781
Tested-by: Jenkins
Reviewed-by: Andreas Heinisch <andreas.heinisch@yahoo.de>