sw: introduce a .uno:GotoMark command
This is meant to be used from macros, it does the same as manually
opening the navigator, selecting a given bookmark and pressing enter.

It also helps debugging when you want to jump somewhere, but the
surrounding document content changes, so jumping to a certain page won't
work due to the changing page number.

Change-Id: I3bc08ae8023997bab89cf4732fb67233a6a1e134
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/143825
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins