crashtesting: (assert) fix treatment of "#FMT" input line value
For "#FMT" test the original format type for not being NUMBER, the current
format type may be different from finding an edit format.

This constellation did not matter until

    commit 4fd1333ba4bb4f2311e9098291154772bd310429
    CommitDate: Thu Mar 25 15:00:31 2021 +0100

        tdf#140968 tdf#140978 XLSX import: fix lost rounded filters

introduced the bFiltering parameter in which case the edit format is not
used (and bFiltering is also set when collecting the filter entries for
.xls export). In fact the logic should be changed such that obtaining
the edit format isn't even executed in that case. For now just fix the
"#FMT" case to be backported.

Also, the bool return of
SvNumberformat::GetOutputString(double,OUString&,Color**) does not indicate
success or failure, but whether the "star" asterisk spreading was inserted or
not. Contrary to SvNumberformat::GetOutputString(double,sal_uInt16,OUString&)
... cough.

Change-Id: Ic8cbbd283a80c654a9ff22ea36897c9b72b8837d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144110
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins