sw: add a new .uno:UpdateBookmarks UNO command
Current the .uno:InsertBookmark command allows inserting a bookmark with
a provided content into the document, but an existing bookmark can't be
updated similarly. This is a problem in case Zotero citations are to be
modeled with bookmarks.

Another trouble is that bookmarks don't have dummy characters, so we
need to be careful to replace the content in a way that maintains these
bookmarks, a naive delete + insert will collapse them.

Fix the problem by introducing a new .uno:UpdateBookmarks command,
somewhat similar to what commit 7765b442e13048f857fd7ee49ced1731caee297e
(sw: add a new .uno:TextFormFields UNO command, 2022-11-28) did.

As usual, the provided new text is meant to be HTML, which allows
multi-paragraph, formatted content.

Change-Id: Id5bb3fd2fd9e147d98709b5a7e0c322e73c6d283
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144167
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins