tdf#152533: Improve script handling in font preview
Use Edit Engine to get the script types instead of trying to duplicate
(poorly) its behaviour. This has the advantage of handling weak
characters and digits properly and is closer to what would happen in the
document itself.

We probably should go further and render the preview entirely using Edit
Engine (since the current code fails short of proper bidi handling), but
this is a bigger change.

Change-Id: I975cb2d96a4a18dbd8110686ca09649cab0ec2f1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144284
Tested-by: Jenkins
Reviewed-by: خالد حسني <khaled@aliftype.com>