tdf#140666 Make delete reference possible from the Navigator
This patch makes it possible to delete a reference without having to do
the steps given in the help.

"References are fields. To remove a reference, delete the field. If you
set a longer text as a reference and you do not want to re-enter it
after deleting the reference, select the text and copy it to the
clipboard. You can then reinsert it as "unformatted text" at the same
position using the command Edit - Paste special. The text remains
intact while the reference is deleted."

Change-Id: If363340288de25aacbc28acd5858ce275f309af6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144406
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>