tdf#42437 Enable press-and-hold special character input method
The first step to enable this input method is to always return a valid selected
range location. -[NSResponder interpretKeyEvents:] will not call
[self firstRectForCharacterRange:actualRange:] if the selected range location
is NSNotFound so that is what was disabling display of the press-and-hold
input method popup.

Once [self firstRectForCharacterRange:actualRange:] is called, emulate the
press-and-hold behavior of the TextEdit application by setting the marked text
to the last key down event's characters. The characters will already have been
committed by the special character input method so set the
mbTextInputWantsNonRepeatKeyDown flag to indicate that the characters need to
be deleted if the input method replaces the committed characters.

Also, emulate the press-and-hold behavior of the TextEdit application by
committing any uncommitted text when either the escape or the return key is
pressed while the press-and-hold input method popup is displayed.

Lastly, handle repeat key events by explicitly inserting the text if there is
no uncommitted text before -[NSResponder interpretKeyEvents:] is called.

Change-Id: I104b3d64a8c66fef3a5a1cec0fba1daa553cd9a4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144630
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>