tdf#152535 Provide better localization for "Grand" subtotal strings
As discussed in the bug report, having the word "Grand" hardcoded first in subtotals is not great for translation, since for some languages the translation of "grand" would come second.

This patch creates separate strings for each case of Grand, f.i. "Grand Count", "Grand Sum" and so on. Now each of them will have their own translation.

Change-Id: Ib875bc1a7b2fcc8934ad85bfe09374f8d9d9a179
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144666
Tested-by: Jenkins
Reviewed-by: Olivier Hallot <olivier.hallot@libreoffice.org>