Corrected errors when Blur effect was re-used in shadow
For GlowPrimitive2D and SoftEdgePrimitive2D I had to correct
the view-dependent decomposition in the case where shadow
was activated for the shape that was used. That creates an
extra transformation, e.g. shadow offset, that needs to be
taken into account, here by operating in discrete view-
coordinates to solve the problem.

NOTE: This is not needed for ShadowPrimitive2D itself when
ShadowBlur is used due to the primitives this is based on
are already handled with the needed offset.

Change-Id: I87dd8db9467fbd27008d6154b78599f28057ce45
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144693
Tested-by: Jenkins
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>