tdf#151704 sw: don't give random height to text frame
This was introduced forever ago with
commit 6c3ae34e32539f8493a940666dbe16b23a8ba7b0
Author: Frank Meies on Tue Nov 20 15:24:54 2001 +0000
    Chg: Vertical Formatting - Growing frames

But why?
Assuming that anything that NEEDED to set a proper height
has done so by now. The commit suggests it was added to
handle vertical layouts.

If this exploratory patch causes problems
(and it very well might since this is a really generic spot)
then perhaps it can be limited to verical layout situations?

Change-Id: Ib6e4a45379e670fd343a2e2d87879e6bb52afebf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144787
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>