sw UI: fix crash on inserting a fieldmark inside a fieldmark
Inserting a fieldmark using .uno:TextFormField, then entering into that
fieldmark using the cursor, finally doing in insertion again using
.uno:TextFormField resulted in a crash.

The problem is that lcl_SetFieldMarks() uses 3
IDocumentContentOperations().InsertString() calls to insert the field
start/separator/end, but right after inserting the field start we
already create an sw::InsertText hint, which works with an inconsistent
string (the start is already inserted but not the separator / end).

Fix the problem by just not allowing the insertion of fieldmarks inside
fieldmarks on the UI: these are meant to be read-only for the user, so
fieldmark insertion is OK to be not working, as long as a clear error
message is provided.

An alternative approach would be to insert the inner fieldmark in a way
similar to how import filters can do it, but that would be more work.

Change-Id: I7d1a7c638b179fd9149ccdc47215329e3433b6e5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144947
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins